+++ lib/src/chart/radar_chart/radar_chart.dart
@@ -1,3 +1,5 @@
+import 'package:fl_chart/src/chart/base/base_chart/initial_animation_configuration.dart';
+import 'package:fl_chart/src/chart/base/base_chart/initial_animation_mixin.dart';
 import 'package:fl_chart/src/chart/radar_chart/radar_chart_data.dart';
 import 'package:fl_chart/src/chart/radar_chart/radar_chart_renderer.dart';
 import 'package:flutter/material.dart';
@@ -11,6 +13,8 @@ class RadarChart extends ImplicitlyAnimatedWidget {
   /// which default is [Curves.linear].
   const RadarChart(
     this.data, {
+    this.chartRendererKey,
+    this.initialAnimationConfiguration = const InitialAnimationConfiguration(),
     super.key,
     Duration swapAnimationDuration = const Duration(milliseconds: 150),
     Curve swapAnimationCurve = Curves.linear,
@@ -22,36 +26,59 @@ class RadarChart extends ImplicitlyAnimatedWidget {
   /// Determines how the [RadarChart] should be look like.
   final RadarChartData data;
 
+  /// We pass this key to our renderers which are supposed to
+  /// render the chart itself (without anything around the chart).
+  final Key? chartRendererKey;
+
+  /// Determines if the initial animation is enabled.
+  final InitialAnimationConfiguration initialAnimationConfiguration;
+
   @override
   _RadarChartState createState() => _RadarChartState();
 }
 
-class _RadarChartState extends AnimatedWidgetBaseState<RadarChart> {
+class _RadarChartState extends AnimatedWidgetBaseState<RadarChart>
+    with InitialAnimationMixin<RadarChartData, RadarChart> {
   /// we handle under the hood animations (implicit animations) via this tween,
   /// it lerps between the old [RadarChartData] to the new one.
   RadarChartDataTween? _radarChartDataTween;
 
   @override
+  InitialAnimationConfiguration get initialAnimationConfiguration =>
+      widget.initialAnimationConfiguration;
+
+  @override
+  Tween? get tween => _radarChartDataTween;
+
+  @override
   Widget build(BuildContext context) {
-    final showingData = _getDate();
+    final showingData = _getData();
 
     return RadarChartLeaf(
+      key: widget.chartRendererKey,
       data: _radarChartDataTween!.evaluate(animation),
       targetData: showingData,
     );
   }
 
-  RadarChartData _getDate() {
+  RadarChartData _getData() {
     return widget.data;
   }
 
   @override
+  RadarChartData getAppearanceAnimationData(RadarChartData data) {
+    return data.copyWith(scaleFactor: 0);
+  }
+
+  @override
   void forEachTween(TweenVisitor<dynamic> visitor) {
-    _radarChartDataTween = visitor(
-      _radarChartDataTween,
-      widget.data,
-      (dynamic value) =>
-          RadarChartDataTween(begin: value as RadarChartData, end: widget.data),
-    ) as RadarChartDataTween?;
+    _radarChartDataTween =
+        visitor(_radarChartDataTween, _getData(), (dynamic value) {
+      final initialData = constructInitialData(value as RadarChartData);
+      return RadarChartDataTween(
+        begin: initialData,
+        end: initialData,
+      );
+    }) as RadarChartDataTween?;
   }
 }