+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## nextVersion
+* **FEATURE** (by @Dartek12) Implemented animations on start for LineChart, BarChart and RadarChart, #406
+
 ## 0.64.0
 * **BUGFIX** (by @Anas35) Fix Tooltip not displaying when value from BackgroundBarChartRodData is less than zero. #1345.
 * **BUGFIX** (by @imaNNeo) Fix Negative BarChartRodStackItem are not drawn correctly bug, #1347